Unified VariableList tests
